% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crime.R
\docType{data}
\name{Crime}
\alias{Crime}
\title{Data from the US FBI Uniform Crime Report, 1960}
\format{A \code{data.frame} \code{Crime}}
\source{
FBI Uniform Crime Report via \href{http://dasl.datadesk.com/data/view/114}{DASL: Data and Story Library}
}
\usage{
data(Crime, package = "statisticalModeling")
}
\description{
A report of the number of offenses reported to police per million
population, and many other social and demographic variables. Each
case corresponds to a state in the US.
}
\examples{
mod_1 <- lm(R ~ W, data = Crime)
mod_2 <- lm(R ~ X, data = Crime)
mod_3 <- lm(R ~ W + X, data = Crime)
effect_size(mod_1, ~ W)
effect_size(mod_3, ~ W)
effect_size(mod_2, ~ X)
effect_size(mod_3, ~ X)
}
\keyword{datasets}

