% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold.R
\name{cv_pred_error}
\alias{cv_pred_error}
\title{Compare models with k-fold cross-validation}
\usage{
cv_pred_error(..., k = 10, ntrials = 5, output = c("mse", "likelihood",
  "error_rate", "class"))
}
\description{
Compare models with k-fold cross-validation
}

