% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_access.R
\name{data_from_model}
\alias{data_from_model}
\alias{explanatory_vars}
\alias{response_var}
\title{Extract component parts of models}
\usage{
data_from_model(object, ...)

explanatory_vars(object, ...)

response_var(object, ...)
}
\arguments{
\item{object}{the model you are extracting features from.}

\item{...}{additional arguments (not used)}
}
\description{
Functions for extracting data and names of explanatory and response variables from a model object
}

