% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{fframe}
\alias{fbar}
\alias{fboxplot}
\alias{fcounts}
\alias{fdensity}
\alias{fframe}
\alias{fhistogram}
\alias{fline}
\alias{fpoint}
\title{Shortcuts for formula_gg}
\usage{
fframe(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

fpoint(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

fline(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

fdensity(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

fboxplot(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

fhistogram(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

fcounts(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

fbar(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)
}
\arguments{
\item{data}{A data frame with the variables to be plotted}

\item{formula}{A formula describing the x and y (if any) variables and other aesthetics in
a form like \code{y ~ x + color:red + shape:sex + alpha:0.5}}

\item{add}{If \code{TRUE} then construct just the layer with no frame.  The result
can be added to an existing frame.}

\item{verbose}{If \code{TRUE} print the ggplot2 command in the console.}

\item{system}{Which graphics system to use, e.g. ggplot2, and so on.}

\item{...}{Other arguments such as \code{position="dodge"},}
}
\description{
These are intended for user convenience, just a small set of plots.
}

