% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmaps.R
\name{fmap}
\alias{fmap}
\title{Drawing a data map underlayed by a map type}
\usage{
fmap(data = NULL, formula = NULL, map = NULL, extent = c("panel",
  "device", "normal"), location = "Saint Paul, MN", zoom = 10,
  verbose = TRUE, source = c("google", "stamen"), type = c("roadmap",
  "terrain", "satellite", "hybrid", "watercolor", "toner"), ...)
}
\arguments{
\item{data}{data to use in drawing glyphs on the map}

\item{formula}{a formula specifying which variables from data are to be latitude ~ longitude}

\item{map}{one way to specify a map}

\item{extent}{the way to handle the frame when overplotting glyphs on the map. See \code{ggmap()}}

\item{location}{another way to specify a map: name of a country, city, etc.}

\item{zoom}{if using \code{location}, then set the scale of the map.
3 = continent, 18 = city block. Each additional step zooms in by a factor of 2 in linear dim.}

\item{verbose}{whether to print the ggplot2 statement (default: \code{TRUE})}

\item{source}{if using \code{location}, the source of the map tiles: Google or Stamen}

\item{type}{if using \code{location}, the style of the map tiles}

\item{...}{additional arguments to be passed to the layer overlaying the map}
}
\description{
This function creates a ggplot2 frame. 
You can add other layers on top of them but you cannot add this frame to an existing one. (Nor should you need to.)
}
\details{
The styles are, for google, \code{"roadmap", "terrain", "hybrid", "satellite"},
for stamen, \code{"terrain", "toner", "watercolor"}.
}
\examples{
\dontrun{
world <- ggmap::get_map(location = 'World', zoom = 3,
                        source = 'google', maptype = 'roadmap', crop=FALSE)
data(CountryCentroids, package = "DataComputing")
fmap(CountryCentroids, lat ~ long, map = world)
}
}

