% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fmodel}
\alias{fmodel}
\title{Plot out model values}
\usage{
fmodel(model = NULL, formula = NULL, data = NULL, nlevels = 3, ...)
}
\arguments{
\item{model}{the model to display graphically}

\item{formula}{setting the y ~ x + color variables}

\item{data}{optional data set from which to extract levels for explanatory variables}

\item{nlevels}{how many levels to display for those variables shown at discrete levels}

\item{...}{arguments to predict()}
}
\description{
Plot out model values
}

