% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggformula.R
\name{formula_to_gg}
\alias{formula_gg}
\alias{formula_to_gg}
\title{Formula interface to ggplot2}
\usage{
formula_to_gg(data = NULL, formula = NULL, add = FALSE, geom = NULL,
  ...)

formula_gg(data = NULL, formula = NULL, add = FALSE, geom = NULL, ...)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula constructed from the variables in the data frame}

\item{add}{if TRUE, create just the layer}

\item{geom}{the name of the geom, e.g. "point", "line", "path". You don't need to use this.}

\item{...}{additional arguments for the ggplot2 geom, e.g. \code{bins = 10} for histogram}
}
\value{
A character string containing the ggplot2 command equivalent of the formula

The ggplot object itself
}
\description{
Provides a model formula interface to ggplot2
}
\details{
Use \code{formula_gg()} to create the plot. \code{formula_to_gg()} translates from
formula notation to an equivalent ggplot2 command.

The formula will, generically, look like
y ~ x for 2-variable graphics, and
~ x for 1-variable graphics (e.g. density, histogram)
For both 2- and 1-variable graphics, you can add terms to the formula to handle other aesthetics
y ~ x + color:blue + shape:age + alpha:0.5
Names like <blue> that are not variables in the data frame will be
set to the corresponding attribute.
Variables and words (e.g. <blue> in color:blue) should **not** be quoted
in the formulas

Some geoms take additional arguments, like \code{bins = } in histograms.  Add those arguments
as you would in the geom.
}
\examples{
data(CPS85, package = "mosaicData")
require(ggplot2)
formula_gg(wage ~ age, data = CPS85)
formula_gg(wage ~ age + color:sex + size:3, data = CPS85) + facet_wrap(~ sector)
formula_gg(wage ~ age, data = CPS85, position = "jitter")
formula_gg( ~ age + color:married + fill:married + alpha:0.5, data = CPS85,
              geom = "histogram", bins = 10)
}

