% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrings.R
\name{frame_string}
\alias{frame_string}
\title{Functions for constructing ggplot command strings}
\usage{
frame_string(data = NULL, x = NULL, y = NULL)
}
\description{
Functions for constructing ggplot command strings
}

