% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{gf_frame}
\alias{gf_bar}
\alias{gf_boxplot}
\alias{gf_counts}
\alias{gf_density}
\alias{gf_frame}
\alias{gf_histogram}
\alias{gf_line}
\alias{gf_point}
\title{Shortcuts for formula_gg}
\usage{
gf_frame(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

gf_point(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

gf_line(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

gf_density(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

gf_boxplot(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

gf_histogram(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

gf_counts(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)

gf_bar(data = NULL, formula = NULL, add = FALSE, verbose = TRUE,
  system = "ggplot2", ...)
}
\arguments{
\item{data}{A data frame with the variables to be plotted}

\item{formula}{A formula describing the x and y (if any) variables and other aesthetics in
a form like \code{y ~ x + color:red + shape:sex + alpha:0.5}}

\item{add}{If \code{TRUE} then construct just the layer with no frame.  The result
can be added to an existing frame.}

\item{verbose}{If \code{TRUE} print the ggplot2 command in the console.}

\item{system}{Which graphics system to use, e.g. ggplot2, and so on.}

\item{...}{Other arguments such as \code{position="dodge"},}
}
\description{
These are intended for user convenience, just a small set of plots.
}

