% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_functions.R
\name{gf_counts}
\alias{gf_bar}
\alias{gf_counts}
\alias{gf_dens}
\alias{gf_density}
\alias{gf_dotplot}
\alias{gf_freqpoly}
\alias{gf_histogram}
\alias{gf_qq}
\title{Univariate gf_ plotting functions}
\usage{
gf_counts(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, debug = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_bar(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, debug = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_freqpoly(placeholder = NULL, formula = NULL, data = NULL,
  geom = type, verbose = FALSE, debug = FALSE,
  add = inherits(placeholder, c("gg", "ggplot")), ...)

gf_density(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, debug = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_histogram(placeholder = NULL, formula = NULL, data = NULL,
  geom = type, verbose = FALSE, debug = FALSE,
  add = inherits(placeholder, c("gg", "ggplot")), ...)

gf_dotplot(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, debug = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_dens(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, debug = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_qq(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, debug = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)
}
\arguments{
\item{placeholder}{Ignore this argument. See details.}

\item{formula}{A formula describing the x variable and other aesthetics in
a form like \code{ ~ x + color:red + fill:gray50 + alpha:0.5}}

\item{data}{A data frame with the variables to be plotted}

\item{geom}{A way to specify ggplot geoms that are not aliased to gf functions.}

\item{verbose}{If \code{TRUE} print the ggplot2 command in the console.}

\item{add}{If \code{TRUE} then construct just the layer with no frame.  The result
can be added to an existing frame.}

\item{...}{Other arguments such as \code{position="dodge"}.}
}
\description{
These functions provide a formula interface to \code{ggplot2} and
various geoms. For plots with just one layer, the formula interface
is more compact and is consistent with modeling and mosaic notation.
The functions generate a \code{ggplot} command string which can be displayed by
setting \code{verbose = TRUE} as an argument.
}
\details{
These \code{gf_} functions are written to interact with ggplot objects.
The \code{placeholder}
argument is part of this interaction system; the end user can ignore it.
}
\examples{
gf_histogram(~ Sepal.Length + fill:Species, data = iris)
gf_density(~ Sepal.Length + color:Species, data = iris)
gf_dens(~ Sepal.Length + color:Species, data = iris)
gf_freqpoly(~ Sepal.Length + color:Species, data = iris)
gf_dotplot(~ Sepal.Length + fill:Species, data = iris)
gf_counts(~ substance, data = HELPrct)
}
\seealso{
\code{\link{gf_point}()}, \code{\link{gf_abline}()}, \code{\link{gf_pointrange}()}.
}

