% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_function_helpers.R
\name{gf_curve}
\alias{gf_crossbar}
\alias{gf_curve}
\alias{gf_errorbar}
\alias{gf_errorbarh}
\alias{gf_linerange}
\alias{gf_pointrange}
\alias{gf_rect}
\alias{gf_ribbon}
\alias{gf_segment}
\title{Multivariate gf_ plotting functions}
\usage{
gf_curve(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(placeholder, c("gg", "ggplot")), ...)

gf_segment(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(placeholder, c("gg", "ggplot")), ...)

gf_ribbon(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(placeholder, c("gg", "ggplot")), ...)

gf_linerange(placeholder = NULL, formula = NULL, data = NULL,
  geom = type, verbose = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_pointrange(placeholder = NULL, formula = NULL, data = NULL,
  geom = type, verbose = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_crossbar(placeholder = NULL, formula = NULL, data = NULL,
  geom = type, verbose = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_errorbar(placeholder = NULL, formula = NULL, data = NULL,
  geom = type, verbose = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_errorbarh(placeholder = NULL, formula = NULL, data = NULL,
  geom = type, verbose = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_rect(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(placeholder, c("gg", "ggplot")), ...)
}
\arguments{
\item{placeholder}{Ignore this argument. See details.}

\item{formula}{A formula describing the manditory aesthetics and possibly other
aesthetics in a form like \code{ y + ymin + ymax ~ x + color:red + fill:gray50 + alpha:0.5}}

\item{data}{A data frame with the variables to be plotted}

\item{geom}{A way to specify ggplot geoms that are not aliased to gf functions.}

\item{verbose}{If \code{TRUE} print the ggplot2 command in the console.}

\item{add}{If \code{TRUE} then construct just the layer with no frame.  The result
can be added to an existing frame.}

\item{...}{Other arguments such as \code{position="dodge"}.}
}
\description{
These functions provide a formula interface to \code{ggplot2} and
various geoms. For plots with just one layer, the formula interface
is more compact and is consistent with modeling and mosaic notation.
The functions generate a \code{ggplot} command string which can be displayed by
setting \code{verbose = TRUE} as an argument.
}

