% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggformula.R
\name{glineg}
\alias{gdensityg}
\alias{gfreqpolyg}
\alias{ghistogramg}
\alias{glineg}
\alias{gpathg}
\alias{gpointg}
\title{Formula interface to ggplot2}
\usage{
glineg(data = parent.env(), formula = NULL, add = FALSE, geom = "line",
  ...)

gpathg(data = parent.env(), formula = NULL, add = FALSE, geom = "path",
  ...)

gpointg(data = parent.env(), formula = NULL, add = FALSE,
  geom = "point", ...)

gfreqpolyg(data = parent.env(), formula = NULL, add = FALSE, ...)

ghistogramg(data = parent.env(), formula = NULL, add = FALSE, ...)

gdensityg(data = parent.env(), formula = NULL, add = FALSE,
  geom = "density", ...)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula constructed from the variables in the data frame}

\item{add}{if TRUE, create just the layer}

\item{geom}{the name of the geom, e.g. "point", "line", "path". You don't need to use this.}
}
\description{
Uses a model formula interface to map and set aesthetics in ggplot2
}
\details{
The formula will, generically, look like
y ~ x
You can add in terms to handle other aesthetics
y ~ x + color:blue + shape:age + alpha:0.5
Names like <blue> that are not variables in the data frame will be
set to the corresponding attribute.
Variables and words (e.g. <blue> in color:blue) should **not** be quoted
in the formulas

Some geoms take additional arguments, like \code{bins = } in histograms.  Add those arguments
as you would in the geom.
}
\examples{
data(CPS85, package = "mosaicData")
gpointg(wage ~ age, data = CPS85)
gpointg(wage ~ age + color:sex + size:3, data = CPS85) + facet_wrap(~ sector)
gpointg(wage ~ age, data = CPS85, position = '"jitter"') # note extra ' ' for character strings

ghistogramg( ~ wage + color:sex, data = CPS85, bins = 5, position="dodge")
# also, position = "fill" for conditional probabilities.
}

