% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold.R
\name{kfold_trial}
\alias{kfold_trial}
\title{returns a vector of predictions or likelihoods}
\usage{
kfold_trial(mod, k = 10, type = c("response", "prob", "likelihood",
  "class"))
}
\description{
returns a vector of predictions or likelihoods
}

