% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggformula.R
\name{line_plot}
\alias{line_plot}
\title{Formula interface to ggplot2}
\usage{
line_plot(data = parent.env(), formula = NULL, add = FALSE,
  geom = "line")
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula constructed from the variables in the data frame}

\item{add}{if TRUE, create just the layer}

\item{geom}{the name of the geom, e.g. "point", "line", "path"}
}
\description{
Formula interface to ggplot2
}

