% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_access.R
\name{reference_values}
\alias{reference_values}
\title{Compute sensible values from a data set for use as a baseline}
\usage{
reference_values(data, n = 1, at = list())
}
\arguments{
\item{data}{a data frame}

\item{n}{number of values for specified variables: could be a single number or
a list assigning a number of levels for individual variables}

\item{at}{optional values at which to set values: a list whose names are the
variables whose values are to be set.}
}
\description{
Compute sensible values from a data set for use as a baseline
}
\details{
Variables not listed in \code{at} will be assigned levels using these principles:
Categorical variables: the most populated levels.
Quantitative variables: central quantiles, e.g. median for n=1, 0.33 and 0.67 for n=2, and so on.
}

