% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Training statistical models}
\usage{
train(data, formula = formula(data), architecture = "lm")
}
\arguments{
\item{data}{the data to use for training}

\item{formula}{the formula describing the structure of the relationship among variables}

\item{architecture}{the model architecture, given as a function name, quoted or not, e.g. lm or "lm"}
}
\description{
This function coordinates four components in building statistical models:
1) the data to use for training
2) the response variable and
3) the explanatory variables, specified as a model formula
4) the architecture of the model.
}
\details{
This is a light wrapper on the architecture-specific function fitting programs
}

