% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{effect_size}
\alias{effect_size}
\title{Calculate effect sizes in a model}
\usage{
effect_size(model, formula, at = NULL, step = NULL, raw = FALSE, ...)
}
\arguments{
\item{model}{the model from which the effect size is to be calculated}

\item{formula}{a formula whose right-hand side is the variable with respect
to which the effect size is to be calculated.}

\item{at}{center values for evaluating the model}

\item{step}{list of numbers for numerical finite-difference, or comparison groups for}

\item{raw}{for 2nd-order effects, show the sets of first-order effects}

\item{...}{additional arguments for \code{predict()}
categorical variables.}
}
\description{
Like a derivative or finite-difference
}
\details{
If two different variables are listed in \code{formula}, the second-order (interaction) effect size
will be calculated.  To Do? Allow 2nd deriv w.r.t a single variable? Get multiple effect sizes all at once.
}

