% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size}
\alias{effect_size}
\title{Calculate effect sizes in a model}
\usage{
effect_size(model, formula, from = NULL, at = NULL, step = NULL,
  to = step, data = NULL, ...)
}
\arguments{
\item{model}{the model from which the effect size is to be calculated}

\item{formula}{a formula whose right-hand side is the variable with respect
to which the effect size is to be calculated.}

\item{from}{optional: the value to use as baseline for the variable specified
in the formula}

\item{at}{center values for evaluating the model}

\item{step}{the numerical stepsize for the change var, or a comparison category
for a categorical change var.}

\item{...}{additional arguments for \code{predict()}. For instance, for a glm, perhaps you
want \code{type = "response"}.}
}
\description{
Like a derivative or finite-difference
}
\examples{
mod1 <- lm(wage ~ age * sex * educ + sector, data = mosaicData::CPS85)
effect_size(mod1, ~ sex)
effect_size(mod1, ~ sector)
effect_size(mod1, ~ age, at = list(sex = "M"))
effect_size(mod1, ~ age, at = list(sex = "F"))
}

