% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fmodel}
\alias{fmodel}
\title{Plot out model values}
\usage{
fmodel(model = NULL, formula = NULL, data = NULL, nlevels = 3, ...)
}
\arguments{
\item{...}{arguments to predict()}
}
\description{
Plot out model values
}

