% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_functions.R
\name{add_arg_list_to_function_string}
\alias{add_arg_list_to_function_string}
\alias{gf_abline}
\alias{gf_bar}
\alias{gf_boxplot}
\alias{gf_counts}
\alias{gf_density}
\alias{gf_density_2d}
\alias{gf_freqpoly}
\alias{gf_hex}
\alias{gf_histogram}
\alias{gf_hline}
\alias{gf_jitter}
\alias{gf_line}
\alias{gf_path}
\alias{gf_point}
\alias{gf_text}
\title{gf_ plotting functions}
\usage{
add_arg_list_to_function_string(S, extras)

gf_point(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_jitter(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_line(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_path(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_density(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_density_2d(placeholder = NULL, formula = NULL, data = NULL,
  geom = type, verbose = FALSE, add = FALSE, ...)

gf_hex(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_hline(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_abline(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_hex(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_boxplot(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_freqpoly(placeholder = NULL, formula = NULL, data = NULL,
  geom = type, verbose = FALSE, add = FALSE, ...)

gf_histogram(placeholder = NULL, formula = NULL, data = NULL,
  geom = type, verbose = FALSE, add = FALSE, ...)

gf_text(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_counts(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)

gf_bar(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = FALSE, ...)
}
\arguments{
\item{formula}{A formula describing the x and y (if any) variables and other aesthetics in
a form like \code{y ~ x + color:red + shape:sex + alpha:0.5}}

\item{data}{A data frame with the variables to be plotted}

\item{verbose}{If \code{TRUE} print the ggplot2 command in the console.}

\item{add}{If \code{TRUE} then construct just the layer with no frame.  The result
can be added to an existing frame.}

\item{...}{Other arguments such as \code{position="dodge"},}

\item{system}{Which graphics system to use, e.g. ggplot2, and so on.}
}
\description{
These functions provide a formula interface to \code{ggplot2} and 
various geoms. For plots with just one layer, the formula interface
is more compact and is consistent with modeling and mosaic notation.
The functions generate a \code{ggplot} command string. The \code{gf_geom_str()} 
versions return that string, while the \code{gf_geom()} return the plot object itself.
}

