% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_functions.R
\name{gf_hline}
\alias{gf_abline}
\alias{gf_coefline}
\alias{gf_hline}
\alias{gf_vline}
\title{gf_ functions with no formula part}
\usage{
gf_hline(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, debug = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_vline(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, debug = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_abline(placeholder = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, debug = FALSE, add = inherits(placeholder, c("gg",
  "ggplot")), ...)

gf_coefline(placeholder = NULL, formula = NULL, coef, ...)
}
\arguments{
\item{placeholder}{Ignore this argument. See details.}

\item{formula}{ignored.}

\item{data}{A data frame with the variables to be plotted}

\item{geom}{A way to specify ggplot geoms that are not aliased to gf functions.}

\item{verbose}{If \code{TRUE} print the ggplot2 command in the console.}

\item{add}{If \code{TRUE} then construct just the layer with no frame.  The result
can be added to an existing frame.}

\item{...}{Other arguments such as \code{position="dodge"}.}
}
\description{
These functions provide a formula interface to \code{ggplot2} and
various geoms. For plots with just one layer, the formula interface
is more compact and is consistent with modeling and mosaic notation.
The functions generate a \code{ggplot} command string which can be displayed by
setting \code{verbose = TRUE} as an argument.
}
\seealso{
\code{\link{gf_point}()}, \code{\link{gf_histogram}()}, \code{\link{gf_pointrange}()}
}

